backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
sqlalchemy==2.0.23
aiosqlite==0.19.0
rapidfuzz==3.5.2
numpy==1.26.2
faiss-cpu==1.7.4
sentence-transformers==2.2.2
torch==2.1.1
torchaudio==2.1.1
openai-whisper==20231117
pyaudio==0.2.14
python-dotenv==1.0.0
